:test_tube: Add initial schemathesis test integration
