+++ .prettierignore
@@ -21,4 +21,5 @@ package-lock.json
 poetry.lock
 site/
 stats.html
+templates/
 web/
