+++ pyproject.toml
@@ -78,7 +78,7 @@ pyyaml = "^6.0.0"
 SQLAlchemy = { extras = ["mypy"], version = "^1.4.0" }
 uvicorn = { extras = ["standard"], version = "^0.18.0" }
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 alembic = "^1.8.0"
 bandit = "^1.7.0"
 black = "^22.8.0"
@@ -98,6 +98,7 @@ pytest-mock = "^3.8.0"
 pytest-playwright = "^0.3.0"
 pytest-postgresql = "^4.1.0"
 requests = "^2.28.0"
+schemathesis = "^3.17.0"
 toml = "^0.10.0"
 tox = "^3.26.0"
 types-pyyaml = "^6.0.12"
