+++ tests/util.py
@@ -3,10 +3,13 @@
 
 import json
 import logging
+import os
 from pathlib import Path
 import secrets
 import socket
-from typing import cast, Dict, Optional
+import subprocess
+from subprocess import Popen
+from typing import cast, Dict, Optional, Tuple
 
 from fastapi.testclient import TestClient
 import requests
@@ -44,6 +47,26 @@ def mock_environment(variables: Dict[str, str]) -> Dict[str, str]:
         },
         **variables,
     }
+
+
+def start_server(database: Path) -> Tuple[Popen, str]:
+    """Start server for testing."""
+    port = find_port()
+    url = f"http://localhost:{port}"
+    environment = {"ACRONYMS_DATABASE": f"sqlite+aiosqlite:///{database}"}
+
+    # Running the server via uvicorn directly as a Python function throws
+    # "RuntimeError: asyncio.run() cannot be called from a running event
+    # loop".
+    process = Popen(
+        ["acronyms", "--port", str(port)],
+        env={**os.environ, **mock_environment(environment)},
+        stderr=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+    )
+
+    wait_for_server(url)
+    return process, url
 
 
 def upload_acronyms(
